feat(value-object.ts): add abstract class ValueObject to provide a base class for value objects in the application
The ValueObject class is an abstract class that provides a base implementation for value objects in the application. It takes a generic type parameter `Props` to represent the properties of the value object.

The class has a protected `props` property to store the properties of the value object. The constructor takes the `props` parameter and assigns it to the `props` property.

The class also provides an `equals` method to compare two value objects for equality. It checks if the passed value object is null or undefined, and if its `props` property is undefined. Then it compares the JSON string representation of the `props` property of both value objects for equality.

This class can be extended by other value objects in the application to inherit the equality comparison functionality.