feat(comment-with-author.ts): add CommentWithAuthor value object to represent a comment with author information in the forum domain
The CommentWithAuthor value object is used to encapsulate the properties of a comment with author information in the forum domain. It includes properties such as commentId, content, authorId, author, createdAt, and updatedAt. This value object provides getters for each property and a static create method to create a new instance of CommentWithAuthor.