feat(in-memory-question-comments-repository.ts): add support for retrieving question comments with author information
The `InMemoryQuestionCommentsRepository` now has a new method `findManyByQuestionIdWithAuthor` which allows retrieving question comments along with the author information. This method takes a `questionId` and `page` as parameters and returns a paginated list of `CommentWithAuthor` objects.

To achieve this, a new dependency `InMemoryStudentsRepository` is injected into the `InMemoryQuestionCommentsRepository` constructor. This repository is used to retrieve the author information for each comment.

The method filters the `items` array based on the `questionId`, applies pagination, and maps each comment to a `CommentWithAuthor` object. The author information is retrieved from the `InMemoryStudentsRepository` using the `authorId` of the comment. If the author is not found, an error is thrown.

This change improves the functionality of the `InMemoryQuestionCommentsRepository` by providing a way to retrieve question comments with author information.